requirements.txt
mkdocs
mkdocs-material
pymdown-extensions
pyyaml
